requirements.txt
playwright==1.41.2
python-telegram-bot==20.7